Refactors paper UI (#88627)
## About The Pull Request
No exciting new features added, I only broke up this massive file into a
folder.
## Why It's Good For The Game
Call me react brained but seeing UI files with >500 lines of code just
strikes me as wall of text / sensory overload
## Changelog

N/A
