+++ tgui/packages/tgui/interfaces/PaperSheet/Preview.tsx
@@ -1,99 +1,12 @@
-/**
- * @license MIT
- */
-
-import { clamp } from 'common/math';
-import { classes } from 'common/react';
 import { marked } from 'marked';
-import { Component, createRef, RefObject } from 'react';
+import { Component, RefObject } from 'react';
 
-import { useBackend, useLocalState } from '../backend';
-import { Box, Button, Flex, Section, TextArea } from '../components';
-import { Window } from '../layouts';
-import { sanitizeText } from '../sanitize';
-
-const Z_INDEX_STAMP = 1;
-const Z_INDEX_STAMP_PREVIEW = 2;
-
-const TEXTAREA_INPUT_HEIGHT = 200;
-
-type PaperContext = {
-  // ui_static_data
-  user_name: string;
-  raw_text_input?: PaperInput[];
-  raw_field_input?: FieldInput[];
-  raw_stamp_input?: StampInput[];
-  max_length: number;
-  max_input_field_length: number;
-  paper_color: string;
-  paper_name: string;
-  default_pen_font: string;
-  default_pen_color: string;
-  signature_font: string;
-  sanitize_text: boolean;
-
-  // ui_data
-  held_item_details?: WritingImplement;
-};
-
-type PaperInput = {
-  raw_text: string;
-  font?: string;
-  color?: string;
-  bold?: boolean;
-  advanced_html?: boolean;
-};
-
-type StampInput = {
-  class: string;
-  x: number;
-  y: number;
-  rotation: number;
-};
-
-type FieldInput = {
-  field_index: string;
-  field_data: PaperInput;
-  is_signature: boolean;
-};
-
-type WritingImplement = {
-  interaction_mode: InteractionType;
-  font?: string;
-  color?: string;
-  use_bold?: boolean;
-  stamp_icon_state?: string;
-  stamp_class?: string;
-};
-
-type PaperSheetStamperState = {
-  x: number;
-  y: number;
-  rotation: number;
-  yOffset: number;
-};
-
-type PaperSheetStamperProps = {
-  scrollableRef: RefObject<HTMLDivElement>;
-};
-
-type FieldCreationReturn = {
-  nextCounter: number;
-  text: string;
-};
-
-type StampPosition = {
-  x: number;
-  y: number;
-  rotation: number;
-  yOffset: number;
-};
-
-enum InteractionType {
-  reading = 0,
-  writing = 1,
-  stamping = 2,
-}
+import { useBackend, useLocalState } from '../../backend';
+import { Box, Section } from '../../components';
+import { sanitizeText } from '../../sanitize';
+import { canEdit, tokenizer, walkTokens } from './helpers';
+import { StampView } from './StampView';
+import { FieldInput, InteractionType, PaperContext } from './types';
 
 type PreviewViewProps = {
   scrollableRef: RefObject<HTMLDivElement>;
@@ -101,343 +14,13 @@ type PreviewViewProps = {
   textArea: string;
 };
 
-const canEdit = (heldItemDetails?: WritingImplement): boolean => {
-  if (!heldItemDetails) {
-    return false;
-  }
-
-  return heldItemDetails.interaction_mode === InteractionType.writing;
+type FieldCreationReturn = {
+  nextCounter: number;
+  text: string;
 };
 
 // Regex that finds [____] fields.
 const fieldRegex: RegExp = /\[((?:_+))\]/gi;
-
-// Handles the ghost stamp when attempting to stamp paper sheets.
-class PaperSheetStamper extends Component<PaperSheetStamperProps> {
-  style: null;
-  state: PaperSheetStamperState = { x: 0, y: 0, rotation: 0, yOffset: 0 };
-  scrollableRef: RefObject<HTMLDivElement>;
-
-  constructor(props) {
-    super(props);
-
-    this.style = null;
-    this.scrollableRef = props.scrollableRef;
-  }
-
-  // Stops propagation of a given event.
-  pauseEvent = (e: Event): boolean => {
-    if (e.stopPropagation) {
-      e.stopPropagation();
-    }
-    if (e.preventDefault) {
-      e.preventDefault();
-    }
-    e.cancelBubble = true;
-    e.returnValue = false;
-    return false;
-  };
-
-  handleMouseMove = (e: MouseEvent): void => {
-    const pos = this.findStampPosition(e);
-    if (!pos) {
-      return;
-    }
-
-    this.pauseEvent(e);
-    this.setState({
-      x: pos.x,
-      y: pos.y,
-      rotation: pos.rotation,
-      yOffset: pos.yOffset,
-    });
-  };
-
-  handleMouseClick = (e: MouseEvent): void => {
-    if (e.pageY <= 30) {
-      return;
-    }
-    const { act } = useBackend<PaperContext>();
-
-    act('add_stamp', {
-      x: this.state.x,
-      y: this.state.y + this.state.yOffset,
-      rotation: this.state.rotation,
-    });
-  };
-
-  findStampPosition(e: MouseEvent): StampPosition | void {
-    let rotating;
-    const scrollable = this.scrollableRef.current;
-
-    if (!scrollable) {
-      return;
-    }
-
-    const stampYOffset = scrollable.scrollTop || 0;
-
-    const stamp = document.getElementById('stamp');
-    if (!stamp) {
-      return;
-    }
-
-    if (e.shiftKey) {
-      rotating = true;
-    }
-
-    const stampHeight = stamp.clientHeight;
-    const stampWidth = stamp.clientWidth;
-
-    const currentHeight = rotating ? this.state.y : e.pageY - stampHeight;
-    const currentWidth = rotating ? this.state.x : e.pageX - stampWidth / 2;
-
-    const widthMin = 0;
-    const heightMin = 0;
-
-    const widthMax = scrollable.clientWidth - stampWidth;
-    const heightMax = scrollable.clientHeight - stampHeight;
-
-    const radians = Math.atan2(
-      currentWidth + stampWidth / 2 - e.pageX,
-      currentHeight + stampHeight - e.pageY,
-    );
-
-    const rotate = rotating
-      ? radians * (180 / Math.PI) * -1
-      : this.state.rotation;
-
-    return {
-      x: clamp(currentWidth, widthMin, widthMax),
-      y: clamp(currentHeight, heightMin, heightMax),
-      rotation: rotate,
-      yOffset: stampYOffset,
-    };
-  }
-
-  componentDidMount() {
-    document.addEventListener('mousemove', this.handleMouseMove);
-    document.addEventListener('click', this.handleMouseClick);
-  }
-
-  componentWillUnmount() {
-    document.removeEventListener('mousemove', this.handleMouseMove);
-    document.removeEventListener('click', this.handleMouseClick);
-  }
-
-  render() {
-    const { data } = useBackend<PaperContext>();
-    const { held_item_details } = data;
-
-    if (!held_item_details?.stamp_class) {
-      return;
-    }
-
-    return (
-      <Stamp
-        activeStamp
-        opacity={0.5}
-        sprite={held_item_details.stamp_class}
-        x={this.state.x}
-        y={this.state.y}
-        rotation={this.state.rotation}
-      />
-    );
-  }
-}
-
-// Creates a full stamp div to render the given stamp to the preview.
-export const Stamp = (props) => {
-  const { activeStamp, sprite, x, y, rotation, opacity, yOffset = 0 } = props;
-  const stamp_transform = {
-    left: x + 'px',
-    top: y + yOffset + 'px',
-    transform: 'rotate(' + rotation + 'deg)',
-    opacity: opacity || 1.0,
-    zIndex: activeStamp ? Z_INDEX_STAMP_PREVIEW : Z_INDEX_STAMP,
-  };
-
-  return (
-    <div
-      id="stamp"
-      className={classes(['Paper__Stamp', sprite])}
-      style={stamp_transform}
-    />
-  );
-};
-
-// Overarching component that holds the primary view for papercode.
-export class PrimaryView extends Component {
-  // Reference that gets passed to the <Section> holding the main preview.
-  // Eventually gets filled with a reference to the section's scroll bar
-  // funtionality.
-  scrollableRef: RefObject<HTMLDivElement>;
-
-  // The last recorded distance the scrollbar was from the bottom.
-  // Used to implement "text scrolls up instead of down" behaviour.
-  lastDistanceFromBottom: number;
-
-  // Event handler for the onscroll event. Also gets passed to the <Section>
-  // holding the main preview. Updates lastDistanceFromBottom.
-  onScrollHandler: (this: GlobalEventHandlers, ev: Event) => any;
-
-  constructor(props) {
-    super(props);
-    this.scrollableRef = createRef();
-    this.lastDistanceFromBottom = 0;
-
-    this.onScrollHandler = (ev) => {
-      const scrollable = ev.currentTarget as HTMLDivElement;
-      if (scrollable) {
-        this.lastDistanceFromBottom =
-          scrollable.scrollHeight - scrollable.scrollTop;
-      }
-    };
-  }
-
-  render() {
-    const { act, data } = useBackend<PaperContext>();
-    const {
-      raw_text_input,
-      raw_field_input,
-      default_pen_font,
-      default_pen_color,
-      paper_color,
-      held_item_details,
-      max_length,
-    } = data;
-
-    const useFont = held_item_details?.font || default_pen_font;
-    const useColor = held_item_details?.color || default_pen_color;
-    const useBold = held_item_details?.use_bold || false;
-
-    const [inputFieldData, setInputFieldData] = useLocalState(
-      'inputFieldData',
-      {},
-    );
-
-    const [textAreaText, setTextAreaText] = useLocalState('textAreaText', '');
-
-    const interactMode =
-      held_item_details?.interaction_mode || InteractionType.reading;
-
-    const savableData =
-      textAreaText.length || Object.keys(inputFieldData).length;
-
-    const dmCharacters =
-      raw_text_input?.reduce((lhs: number, rhs: PaperInput) => {
-        return lhs + rhs.raw_text.length;
-      }, 0) || 0;
-
-    const usedCharacters = dmCharacters + textAreaText.length;
-
-    const tooManyCharacters = usedCharacters > max_length;
-
-    return (
-      <>
-        <PaperSheetStamper scrollableRef={this.scrollableRef} />
-        <Flex direction="column" fillPositionedParent>
-          <Flex.Item grow={3} basis={1}>
-            <PreviewView
-              key={`${raw_field_input?.length || 0}_${
-                raw_text_input?.length || 0
-              }`}
-              scrollableRef={this.scrollableRef}
-              handleOnScroll={this.onScrollHandler}
-              textArea={textAreaText}
-            />
-          </Flex.Item>
-          {interactMode === InteractionType.writing && (
-            <Flex.Item shrink={1} height={TEXTAREA_INPUT_HEIGHT + 'px'}>
-              <Section
-                title="Insert Text"
-                fitted
-                fill
-                buttons={
-                  <>
-                    <Box
-                      inline
-                      pr={'5px'}
-                      color={tooManyCharacters ? 'bad' : 'default'}
-                    >
-                      {`${usedCharacters} / ${max_length}`}
-                    </Box>
-                    <Button.Confirm
-                      disabled={!savableData || tooManyCharacters}
-                      content="Save"
-                      color="good"
-                      onClick={() => {
-                        if (textAreaText.length) {
-                          act('add_text', { text: textAreaText });
-                          setTextAreaText('');
-                        }
-                        if (Object.keys(inputFieldData).length) {
-                          act('fill_input_field', {
-                            field_data: inputFieldData,
-                          });
-                          setInputFieldData({});
-                        }
-                      }}
-                    />
-                  </>
-                }
-              >
-                <TextArea
-                  scrollbar
-                  noborder
-                  value={textAreaText}
-                  textColor={useColor}
-                  fontFamily={useFont}
-                  bold={useBold}
-                  height="100%"
-                  backgroundColor={paper_color}
-                  onInput={(e, text) => {
-                    setTextAreaText(text);
-
-                    if (this.scrollableRef.current) {
-                      let thisDistFromBottom =
-                        this.scrollableRef.current.scrollHeight -
-                        this.scrollableRef.current.scrollTop;
-                      this.scrollableRef.current.scrollTop +=
-                        thisDistFromBottom - this.lastDistanceFromBottom;
-                    }
-                  }}
-                />
-              </Section>
-            </Flex.Item>
-          )}
-        </Flex>
-      </>
-    );
-  }
-}
-
-const tokenizer = (src: string) => {
-  const rule = /^\[_+\]/;
-  const match = src.match(rule);
-  if (match) {
-    return {
-      type: 'inputField',
-      raw: match[0],
-    };
-  }
-};
-
-// Override function, any links and images should
-// kill any other marked tokens we don't want here
-const walkTokens = (token) => {
-  switch (token.type) {
-    case 'url':
-    case 'autolink':
-    case 'reflink':
-    case 'link':
-    case 'image':
-      token.type = 'text';
-      // Once asset system is up change to some default image
-      // or rewrite for icon images
-      token.href = '';
-      break;
-  }
-};
 
 /**
  * Real-time text preview section. When not editing, this is simply
@@ -452,7 +35,7 @@ const walkTokens = (token) => {
  */
 export class PreviewView extends Component<PreviewViewProps> {
   // Array containing cache of HTMLInputElements that are enabled.
-  enabledInputFieldCache: { [key: string]: HTMLInputElement } = {};
+  enabledInputFieldCache: Record<string, HTMLInputElement> = {};
 
   // State checking variables. Used to determine whether or not to use cache.
   lastReadOnly: boolean = true;
@@ -955,59 +538,3 @@ export class PreviewView extends Component<PreviewViewProps> {
     );
   }
 }
-
-// Renders all the stamp components for every valid stamp.
-export const StampView = (props) => {
-  const { data } = useBackend<PaperContext>();
-
-  const { raw_stamp_input = [] } = data;
-
-  const { stampYOffset } = props;
-
-  return (
-    <>
-      {raw_stamp_input.map((stamp, index) => {
-        return (
-          <Stamp
-            key={index}
-            x={stamp.x}
-            y={stamp.y}
-            rotation={stamp.rotation}
-            sprite={stamp.class}
-            yOffset={stampYOffset}
-          />
-        );
-      })}
-    </>
-  );
-};
-
-export const PaperSheet = (props) => {
-  const { data } = useBackend<PaperContext>();
-  const { paper_color, paper_name, held_item_details } = data;
-
-  const writeMode = canEdit(held_item_details);
-
-  if (!writeMode) {
-    const [inputFieldData, setInputFieldData] = useLocalState(
-      'inputFieldData',
-      {},
-    );
-    if (Object.keys(inputFieldData).length) {
-      setInputFieldData({});
-    }
-  }
-
-  return (
-    <Window
-      title={paper_name}
-      theme="paper"
-      width={420}
-      height={500 + (writeMode ? TEXTAREA_INPUT_HEIGHT : 0)}
-    >
-      <Window.Content backgroundColor={paper_color}>
-        <PrimaryView />
-      </Window.Content>
-    </Window>
-  );
-};
